requirements.txt
fastapi
uvicorn
pydantic
pydantic-ai
httpx
python-dotenv
ollama